This is the custom camera movement script
There is a Camera Register script that is attached to the cameras in the game and a Camera Switcher script that registers the cameras from the Camera Register Script. There is a Camera Trigger Volume Script attached to empty game objects in the game that are triggered to change cameras when the player walks through them.